+++ modules/item/spell-item.js
@@ -72,10 +72,17 @@ export class HMSpellItem extends HMItem {
 
         const {save} = context.system;
         if (resp.button === 'cast' && save.type > HMCONST.SAVE.TYPE.SPECIAL) {
-            const level = actor.itemTypes.cclass.length
-                ? caller.itemTypes.cclass[0].system.level
-                : parseInt(caller.system.level, 10) || 1;
-            dataset.roll = await new Roll(HMTABLES.formula.save.target, {level})
+            let spellSave = 0;
+            const cclass = actor.itemTypes.cclass[0];
+            if (actor.type === 'beast') spellSave = parseInt(caller.system.level, 10) || 1;
+            if (actor.type === 'character' && cclass) {
+                const {system} = cclass;
+                spellSave = system.features.slvl
+                    ? parseInt(system.bonus.slvl, 10) || 0
+                    : parseInt(system.level, 10) || 1;
+            }
+            console.warn(spellSave);
+            dataset.roll = await new Roll(HMTABLES.formula.save.spell, {spellSave})
                                          .evaluate({async: true});
         }