+++ modules/tables/dictionary.js
@@ -45,19 +45,6 @@ idx.alignment = {
     'non': 'HM.alignmentSelect.non',
 };
 
-idx.classFeatureTable = {
-    assn: 'HM.CCLASS.TABLE.assn',
-    atk:  'HM.CCLASS.TABLE.atk',
-    back: 'HM.CCLASS.TABLE.back',
-    def:  'HM.CCLASS.TABLE.def',
-    init: 'HM.CCLASS.TABLE.init',
-    luck: 'HM.CCLASS.TABLE.luck',
-    sp:   'HM.CCLASS.TABLE.sp',
-    spd:  'HM.CCLASS.TABLE.spd',
-    spdm: 'HM.CCLASS.TABLE.spdm',
-    spdr: 'HM.CCLASS.TABLE.spdr',
-};
-
 idx.containerType = {
     [HMCONST.CONTAINER.TYPE.WEIGHT]:    'HM.containerTypeSelect.weight',
     [HMCONST.CONTAINER.TYPE.QUANTITY]:  'HM.containerTypeSelect.quantity',
@@ -263,21 +250,6 @@ idx.talentSelect = {
     [HMCONST.TALENT.EFFECT]:    'HM.talentSelect.effect',
 };
 
-idx.itemCClass = {
-    sp:    'HM.CCLASS.OPT.sp',
-    init:  'HM.CCLASS.OPT.init',
-    atk:   'HM.CCLASS.OPT.atk',
-    def:   'HM.CCLASS.OPT.def',
-    spd:   'HM.CCLASS.OPT.spd',
-    spdr:  'HM.CCLASS.OPT.spdr',
-    spdm:  'HM.CCLASS.OPT.spdm',
-    back:  'HM.CCLASS.OPT.back',
-    fcast: 'HM.CCLASS.OPT.fcast',
-    hprof: 'HM.CCLASS.OPT.hprof',
-    luck:  'HM.CCLASS.OPT.luck',
-    assn:  'HM.CCLASS.OPT.assn',
-};
-
 idx.itemTypes = {
     armor:     'TYPES.Item.armor',
     weapon:    'TYPES.Item.weapon',
@@ -373,6 +345,32 @@ idx.spellLevels = {
     '19': 'HM.spellLevels.19',
     '20': 'HM.spellLevels.20',
     '21': 'HM.spellLevels.21',
+};
+
+idx.spellHybrid = {
+    0:  'HM.SPELL.HYBRID.0',
+    1:  'HM.SPELL.HYBRID.1',
+    2:  'HM.SPELL.HYBRID.2',
+    3:  'HM.SPELL.HYBRID.3',
+    4:  'HM.SPELL.HYBRID.4',
+    5:  'HM.SPELL.HYBRID.5',
+    6:  'HM.SPELL.HYBRID.6',
+    7:  'HM.SPELL.HYBRID.7',
+    8:  'HM.SPELL.HYBRID.8',
+    9:  'HM.SPELL.HYBRID.9',
+    10: 'HM.SPELL.HYBRID.10',
+    11: 'HM.SPELL.HYBRID.11',
+    12: 'HM.SPELL.HYBRID.12',
+    13: 'HM.SPELL.HYBRID.13',
+    14: 'HM.SPELL.HYBRID.14',
+    15: 'HM.SPELL.HYBRID.15',
+    16: 'HM.SPELL.HYBRID.16',
+    17: 'HM.SPELL.HYBRID.17',
+    18: 'HM.SPELL.HYBRID.18',
+    19: 'HM.SPELL.HYBRID.19',
+    20: 'HM.SPELL.HYBRID.20',
+    21: 'HM.SPELL.HYBRID.21',
+    22: 'HM.SPELL.HYBRID.22',
 };
 
 // Images